refactor(epicenter): rename contract.ts to workspace.ts with enhanced JSDoc
- Rename contract.ts → workspace.ts for clearer naming
- Add module-level JSDoc with ASCII architecture diagram
- Add rich JSDoc to node.ts explaining relationship to workspace.ts
- Derive Node Workspace type from sync version (reduces duplication)
- Update all imports across cli, server, and index files
- Add @see links connecting related modules
